backend/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary>=2.9.9
requests>=2.31.0
certifi>=2023.7.22
charset-normalizer>=3.3.0
click>=8.1.7
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.3
urllib3>=2.0.7
Werkzeug>=3.0.1
python-dotenv==1.0.0
